project1/requirements.txt
flask==2.0.1
pillow==9.5.0
numpy==1.24.3
tensorflow==2.13.0
gunicorn==20.1.0
flask-cors==4.0.0
python-dotenv==1.0.0